requirement.txt
aiohttp==2.0.7
amqp==2.1.4
anyjson==0.3.3
async-timeout==1.2.1
billiard==3.5.0.2
blessings==1.6
bpython==0.16
celery==4.0.2
certifi==2017.4.17
chardet==3.0.3
curtsies==0.2.11
Django==1.11.2
django-bootstrap3==8.2.3
django-celery==3.2.1
django-celery-beat==1.0.1
django-tagging==0.4.5
django-taggit==0.22.1
greenlet==0.4.12
idna==2.5
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
multidict==2.1.5
mysqlclient==1.3.10
olefile==0.44
Pillow==4.1.1
Pygments==2.2.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
six==1.10.0
urllib3==1.21.1
vine==1.1.3
wcwidth==0.1.7
yarl==0.10.2
